Asset_Commentioning_chatbot_backend/requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
sqlite-utils
groq
openai
python-dotenv
pandas